Langchain part -2/requirements.txt
langchain_openai
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette